--- a/benchmark-tests/benches/test_bin_bench/client_requests/expected_stdout.1
+++ b/benchmark-tests/benches/test_bin_bench/client_requests/expected_stdout.1
@@ -1,9 +1,9 @@
 test_bin_bench_client_requests::client_requests::run parse_output:() -> target/release/client-requests
 Hello World.
 - end of stdout/stderr
-  Instructions:                    |N/A             (*********)
-  L1 Hits:                         |N/A             (*********)
-  L2 Hits:                         |N/A             (*********)
-  RAM Hits:                        |N/A             (*********)
-  Total read+write:                |N/A             (*********)
-  Estimated Cycles:                |N/A             (*********)
+  Instructions:                      |N/A             (*********)
+  L1 Hits:                           |N/A             (*********)
+  L2 Hits:                           |N/A             (*********)
+  RAM Hits:                          |N/A             (*********)
+  Total read+write:                  |N/A             (*********)
+  Estimated Cycles:                  |N/A             (*********)