--- a/benchmark-tests/benches/test_bin_bench/command_delay/expected_stdout.1
+++ b/benchmark-tests/benches/test_bin_bench/command_delay/expected_stdout.1
@@ -1,37 +1,37 @@
 test_bin_bench_command_delay::delay::delay_duration delay:() -> target/release/echo
-  Instructions:                    |N/A             (*********)
-  L1 Hits:                         |N/A             (*********)
-  L2 Hits:                         |N/A             (*********)
-  RAM Hits:                        |N/A             (*********)
-  Total read+write:                |N/A             (*********)
-  Estimated Cycles:                |N/A             (*********)
+  Instructions:                      |N/A             (*********)
+  L1 Hits:                           |N/A             (*********)
+  L2 Hits:                           |N/A             (*********)
+  RAM Hits:                          |N/A             (*********)
+  Total read+write:                  |N/A             (*********)
+  Estimated Cycles:                  |N/A             (*********)
 test_bin_bench_command_delay::delay::delay_path delay:() -> target/release/echo
 Waiting to create file...
 Creating file...
 File created...
-  Instructions:                    |N/A             (*********)
-  L1 Hits:                         |N/A             (*********)
-  L2 Hits:                         |N/A             (*********)
-  RAM Hits:                        |N/A             (*********)
-  Total read+write:                |N/A             (*********)
-  Estimated Cycles:                |N/A             (*********)
+  Instructions:                      |N/A             (*********)
+  L1 Hits:                           |N/A             (*********)
+  L2 Hits:                           |N/A             (*********)
+  RAM Hits:                          |N/A             (*********)
+  Total read+write:                  |N/A             (*********)
+  Estimated Cycles:                  |N/A             (*********)
 test_bin_bench_command_delay::delay::delay_tcp delay:() -> target/release/echo
 Waiting to start server...
 Starting server...
 Stopped server...
-  Instructions:                    |N/A             (*********)
-  L1 Hits:                         |N/A             (*********)
-  L2 Hits:                         |N/A             (*********)
-  RAM Hits:                        |N/A             (*********)
-  Total read+write:                |N/A             (*********)
-  Estimated Cycles:                |N/A             (*********)
+  Instructions:                      |N/A             (*********)
+  L1 Hits:                           |N/A             (*********)
+  L2 Hits:                           |N/A             (*********)
+  RAM Hits:                          |N/A             (*********)
+  Total read+write:                  |N/A             (*********)
+  Estimated Cycles:                  |N/A             (*********)
 test_bin_bench_command_delay::delay::delay_udp delay:() -> target/release/echo
 Waiting to start server...
 Starting server...
 Stopped server...
-  Instructions:                    |N/A             (*********)
-  L1 Hits:                         |N/A             (*********)
-  L2 Hits:                         |N/A             (*********)
-  RAM Hits:                        |N/A             (*********)
-  Total read+write:                |N/A             (*********)
-  Estimated Cycles:                |N/A             (*********)
+  Instructions:                      |N/A             (*********)
+  L1 Hits:                           |N/A             (*********)
+  L2 Hits:                           |N/A             (*********)
+  RAM Hits:                          |N/A             (*********)
+  Total read+write:                  |N/A             (*********)
+  Estimated Cycles:                  |N/A             (*********)