--- a/benchmark-tests/benches/test_bin_bench/config/expected_stdout
+++ b/benchmark-tests/benches/test_bin_bench/config/expected_stdout
@@ -5,14 +5,14 @@ Found env: 'BINARY_BENCHMARK_ENV' with value '3'
 Found env: 'BENCH_ENV' with value '4'
 Found env: 'COMMAND_ENV' with value '5'
 - end of stdout/stderr
-  Instructions:                    |N/A             (*********)
-  L1 Hits:                         |N/A             (*********)
-  L2 Hits:                         |N/A             (*********)
-  RAM Hits:                        |N/A             (*********)
-  Total read+write:                |N/A             (*********)
-  Estimated Cycles:                |N/A             (*********)
+  Instructions:                      |N/A             (*********)
+  L1 Hits:                           |N/A             (*********)
+  L2 Hits:                           |N/A             (*********)
+  RAM Hits:                          |N/A             (*********)
+  Total read+write:                  |N/A             (*********)
+  Estimated Cycles:                  |N/A             (*********)
 test_bin_bench_config::my_group::check_env_is_cleared with_env:() -> target/release/env '--is-cleared=true'
-  ======= CALLGRIND =======================================================
+  ======= CALLGRIND =========================================================
 SETUP:
 The environment variables look like they have not been cleared
 Found env: 'MAIN_ENV' with value '1'
@@ -29,13 +29,13 @@ Found env: 'BINARY_BENCHMARK_ENV' with value '3'
 Found env: 'BENCH_ENV' with value '4'
 Found env: 'COMMAND_ENV' with value '5'
 - end of stdout/stderr
-  Instructions:                    |N/A             (*********)
-  L1 Hits:                         |N/A             (*********)
-  L2 Hits:                         |N/A             (*********)
-  RAM Hits:                        |N/A             (*********)
-  Total read+write:                |N/A             (*********)
-  Estimated Cycles:                |N/A             (*********)
-  ======= DHAT ============================================================
+  Instructions:                      |N/A             (*********)
+  L1 Hits:                           |N/A             (*********)
+  L2 Hits:                           |N/A             (*********)
+  RAM Hits:                          |N/A             (*********)
+  Total read+write:                  |N/A             (*********)
+  Estimated Cycles:                  |N/A             (*********)
+  ======= DHAT ==============================================================
 SETUP:
 The environment variables look like they have not been cleared
 Found env: 'MAIN_ENV' with value '1'
@@ -52,12 +52,11 @@ Found env: 'BINARY_BENCHMARK_ENV' with value '3'
 Found env: 'BENCH_ENV' with value '4'
 Found env: 'COMMAND_ENV' with value '5'
 - end of stdout
-  Total bytes:                     |N/A             (*********)
-  Total blocks:                    |N/A             (*********)
-  At t-gmax bytes:                 |N/A             (*********)
-  At t-gmax blocks:                |N/A             (*********)
-  At t-end bytes:                  |N/A             (*********)
-  At t-end blocks:                 |N/A             (*********)
-  Reads bytes:                     |N/A             (*********)
-  Writes bytes:                    |N/A             (*********)
-  Outfile:          target/iai/benchmark-tests/test_bin_bench_config/my_group/check_env_is_cleared.with_env/dhat.check_env_is_cleared.with_env.out
+  Total bytes:                       |N/A             (*********)
+  Total blocks:                      |N/A             (*********)
+  At t-gmax bytes:                   |N/A             (*********)
+  At t-gmax blocks:                  |N/A             (*********)
+  At t-end bytes:                    |N/A             (*********)
+  At t-end blocks:                   |N/A             (*********)
+  Reads bytes:                       |N/A             (*********)
+  Writes bytes:                      |N/A             (*********)