--- a/benchmark-tests/benches/test_bin_bench/exit_with/expected_stdout.1
+++ b/benchmark-tests/benches/test_bin_bench/exit_with/expected_stdout.1
@@ -1,7 +1,7 @@
 test_bench_template::my_group::bench_exit_with exit_with:(ExitWith :: Code(0)) -> target/release/exit-with 0
-  Instructions:                    |N/A             (*********)
-  L1 Hits:                         |N/A             (*********)
-  L2 Hits:                         |N/A             (*********)
-  RAM Hits:                        |N/A             (*********)
-  Total read+write:                |N/A             (*********)
-  Estimated Cycles:                |N/A             (*********)
+  Instructions:                      |N/A             (*********)
+  L1 Hits:                           |N/A             (*********)
+  L2 Hits:                           |N/A             (*********)
+  RAM Hits:                          |N/A             (*********)
+  Total read+write:                  |N/A             (*********)
+  Estimated Cycles:                  |N/A             (*********)