--- a/benchmark-tests/benches/test_bin_bench/exit_with/expected_stdout.2
+++ b/benchmark-tests/benches/test_bin_bench/exit_with/expected_stdout.2
@@ -1,6 +1,6 @@
 test_bench_template::my_group::bench_exit_with exit_with:(ExitWith :: Success) -> target/release/exit-with 0
   Instructions:                      |                (No change)
-  L1 Hits:                           |                (No change)
+  L1 Hits:                           |                (         )
   L2 Hits:                           |                (         )
   RAM Hits:                          |                (         )
   Total read+write:                  |                (No change)