--- a/benchmark-tests/benches/test_bin_bench/exit_with/expected_stdout.4
+++ b/benchmark-tests/benches/test_bin_bench/exit_with/expected_stdout.4
@@ -1,6 +1,6 @@
 test_bench_template::my_group::bench_exit_with exit_with:(ExitWith :: Failure) -> target/release/exit-with 1
   Instructions:                      |                (No change)
-  L1 Hits:                           |                (No change)
+  L1 Hits:                           |                (         )
   L2 Hits:                           |                (         )
   RAM Hits:                          |                (         )
   Total read+write:                  |                (No change)