--- a/benchmark-tests/benches/test_bin_bench/main_and_group_setup_and_teardown/expected_stdout.1
+++ b/benchmark-tests/benches/test_bin_bench/main_and_group_setup_and_teardown/expected_stdout.1
@@ -2,27 +2,27 @@ MAIN SETUP
 GROUP SETUP
 test_bin_bench_main_and_group_setup_and_teardown::simple_group_with_setup::simple_bench () -> target/release/read-file /tmp/iai-callgrind....
 - end of stdout/stderr
-  Instructions:                    |N/A             (*********)
-  L1 Hits:                         |N/A             (*********)
-  L2 Hits:                         |N/A             (*********)
-  RAM Hits:                        |N/A             (*********)
-  Total read+write:                |N/A             (*********)
-  Estimated Cycles:                |N/A             (*********)
+  Instructions:                      |N/A             (*********)
+  L1 Hits:                           |N/A             (*********)
+  L2 Hits:                           |N/A             (*********)
+  RAM Hits:                          |N/A             (*********)
+  Total read+write:                  |N/A             (*********)
+  Estimated Cycles:                  |N/A             (*********)
 test_bin_bench_main_and_group_setup_and_teardown::simple_group_with_setup::check_file_exists () -> target/release/file-exists /tmp/iai-callgrin...
 - end of stdout/stderr
-  Instructions:                    |N/A             (*********)
-  L1 Hits:                         |N/A             (*********)
-  L2 Hits:                         |N/A             (*********)
-  RAM Hits:                        |N/A             (*********)
-  Total read+write:                |N/A             (*********)
-  Estimated Cycles:                |N/A             (*********)
+  Instructions:                      |N/A             (*********)
+  L1 Hits:                           |N/A             (*********)
+  L2 Hits:                           |N/A             (*********)
+  RAM Hits:                          |N/A             (*********)
+  Total read+write:                  |N/A             (*********)
+  Estimated Cycles:                  |N/A             (*********)
 GROUP TEARDOWN
 test_bin_bench_main_and_group_setup_and_teardown::check_group::check_file_not_exists () -> target/release/file-exists /tmp/iai-callgrin...
 - end of stdout/stderr
-  Instructions:                    |N/A             (*********)
-  L1 Hits:                         |N/A             (*********)
-  L2 Hits:                         |N/A             (*********)
-  RAM Hits:                        |N/A             (*********)
-  Total read+write:                |N/A             (*********)
-  Estimated Cycles:                |N/A             (*********)
+  Instructions:                      |N/A             (*********)
+  L1 Hits:                           |N/A             (*********)
+  L2 Hits:                           |N/A             (*********)
+  RAM Hits:                          |N/A             (*********)
+  Total read+write:                  |N/A             (*********)
+  Estimated Cycles:                  |N/A             (*********)
 MAIN TEARDOWN