--- a/benchmark-tests/benches/test_bin_bench/main_and_group_setup_and_teardown/expected_stdout.2
+++ b/benchmark-tests/benches/test_bin_bench/main_and_group_setup_and_teardown/expected_stdout.2
@@ -1,24 +1,24 @@
 test_bin_bench_main_and_group_setup_and_teardown::simple_group_with_setup::simple_bench () -> target/release/read-file /tmp/iai-callgrind....
-  Baselines:                default|default
-  Instructions:                    |                (No change)
-  L1 Hits:                         |                (No change)
-  L2 Hits:                         |                (         )
-  RAM Hits:                        |                (         )
-  Total read+write:                |                (No change)
-  Estimated Cycles:                |                (         )
+  Baselines:                  default|default
+  Instructions:                      |                (No change)
+  L1 Hits:                           |                (No change)
+  L2 Hits:                           |                (         )
+  RAM Hits:                          |                (         )
+  Total read+write:                  |                (No change)
+  Estimated Cycles:                  |                (         )
 test_bin_bench_main_and_group_setup_and_teardown::simple_group_with_setup::check_file_exists () -> target/release/file-exists /tmp/iai-callgrin...
-  Baselines:                default|default
-  Instructions:                    |                (No change)
-  L1 Hits:                         |                (No change)
-  L2 Hits:                         |                (         )
-  RAM Hits:                        |                (         )
-  Total read+write:                |                (No change)
-  Estimated Cycles:                |                (         )
+  Baselines:                  default|default
+  Instructions:                      |                (No change)
+  L1 Hits:                           |                (No change)
+  L2 Hits:                           |                (         )
+  RAM Hits:                          |                (         )
+  Total read+write:                  |                (No change)
+  Estimated Cycles:                  |                (         )
 test_bin_bench_main_and_group_setup_and_teardown::check_group::check_file_not_exists () -> target/release/file-exists /tmp/iai-callgrin...
-  Baselines:                default|default
-  Instructions:                    |                (No change)
-  L1 Hits:                         |                (No change)
-  L2 Hits:                         |                (         )
-  RAM Hits:                        |                (         )
-  Total read+write:                |                (No change)
-  Estimated Cycles:                |                (         )
+  Baselines:                  default|default
+  Instructions:                      |                (No change)
+  L1 Hits:                           |                (No change)
+  L2 Hits:                           |                (         )
+  RAM Hits:                          |                (         )
+  Total read+write:                  |                (No change)
+  Estimated Cycles:                  |                (         )