--- a/benchmark-tests/benches/test_lib_bench/cache_sim/expected_stdout.2
+++ b/benchmark-tests/benches/test_lib_bench/cache_sim/expected_stdout.2
@@ -1,6 +1,6 @@
 test_lib_bench_cache_sim::bench_cache_sim::bench_with_cache_sim with_10:setup_worst_case_array(10)
   Instructions:                      |                (No change)
-  L1 Hits:                           |                (No change)
+  L1 Hits:                           |                (         )
   L2 Hits:                           |                (         )
   RAM Hits:                          |                (         )
   Total read+write:                  |                (No change)