--- a/benchmark-tests/benches/test_lib_bench/setup_and_teardown/test_lib_bench_setup_and_teardown.conf.yml
+++ b/benchmark-tests/benches/test_lib_bench/setup_and_teardown/test_lib_bench_setup_and_teardown.conf.yml
@@ -1,0 +1,3 @@
+groups:
+  - runs:
+      - args: []