--- a/iai-callgrind-runner/tests/test_tool/mod.rs
+++ b/iai-callgrind-runner/tests/test_tool/mod.rs
@@ -1,0 +1,1 @@
+mod test_error_metrics_parser;