This pr implements a better (and correct) handling of the callgrind argument `--separate-threads` and valgrind argument `--trace-children` (#28) which enables Iai-Callgrind to correctly summarize and show the benchmark metrics of multi-threaded and multi-process applications.

To be able to create a total over all threads and subprocesses, Iai-Callgrind needed to be able to parse and summarize multiple output files. Multiple files are usually created by valgrind tools on a per process basis (`--trace-children=yes`). In addition, callgrind separates threads much like subprocesses in a separate file (`--separate-threads=yes`). 

Due to the new processing of multiple output files, the following callgrind arguments creating multiple parts are now supported too: `--dump-every-bb`, `--dump-before`, `--dump-after`.

Per default, Iai-Callgrind now runs all benchmarks with `--trace-children=yes`, `--fair-sched=try` (to produce a more reliable thread execution order on systems which support this scheduler) and callgrind with `--separate-threads=yes`.

The default output format still only shows the total over all threads and subprocesses:

```text
test_lib_bench_threads::bench_group::bench_find_primes_multi_thread two:2
  Instructions:               6140005|6140005         (No change)
  L1 Hits:                    6328020|6328020         (No change)
  L2 Hits:                        394|394             (No change)
  RAM Hits:                       920|920             (No change)
  Total read+write:           6329334|6329334         (No change)
  Estimated Cycles:           6362190|6362190         (No change)
```

This behaviour can be changed with the new `OutputFormat::show_intermediate`, for example in a library benchmark 

```rust
main!(
    config = LibraryBenchmarkConfig::default()
        .output_format(OutputFormat::default()
            .show_intermediate(true)
         );
    library_benchmark_groups = /* ... */
);
```

```text
test_lib_bench_threads::bench_group::bench_find_primes_multi_thread two:2
  ## pid: 564676 part: 1 thread: 1   |pid: 563401 part: 1 thread: 1
  Command:            target/release/deps/test_lib_bench_threads-4c5214398e2f5bd1
  Instructions:                 29084|29084           (No change)
  L1 Hits:                      68782|68782           (No change)
  L2 Hits:                        381|381             (No change)
  RAM Hits:                       603|603             (No change)
  Total read+write:             69766|69766           (No change)
  Estimated Cycles:             91792|91792           (No change)
  ## pid: 564676 part: 1 thread: 2   |pid: 563401 part: 1 thread: 2
  Command:            target/release/deps/test_lib_bench_threads-4c5214398e2f5bd1
  Instructions:               2460507|2460507         (No change)
  L1 Hits:                    2534943|2534943         (No change)
  L2 Hits:                         12|12              (No change)
  RAM Hits:                       187|187             (No change)
  Total read+write:           2535142|2535142         (No change)
  Estimated Cycles:           2541548|2541548         (No change)
  ## pid: 564676 part: 1 thread: 3   |pid: 563401 part: 1 thread: 3
  Command:            target/release/deps/test_lib_bench_threads-4c5214398e2f5bd1
  Instructions:               3650414|3650414         (No change)
  L1 Hits:                    3724295|3724295         (No change)
  L2 Hits:                          1|1               (No change)
  RAM Hits:                       130|130             (No change)
  Total read+write:           3724426|3724426         (No change)
  Estimated Cycles:           3728850|3728850         (No change)
  ## Total
  Instructions:               6140005|6140005         (No change)
  L1 Hits:                    6328020|6328020         (No change)
  L2 Hits:                        394|394             (No change)
  RAM Hits:                       920|920             (No change)
  Total read+write:           6329334|6329334         (No change)
  Estimated Cycles:           6362190|6362190         (No change)
```

Closes #255
Fixes #28